tests: compare WFP bypass weight as uint64

+++ vpn/netstate/routes_windows_test.go
@@ -165,7 +165,7 @@ func TestClientFenceRulesPermitConfiguredBypass(t *testing.T) {
 			continue
 		}
 		require.Equal(t, wf.ActionPermit, r.Action)
-		require.Equal(t, fenceWeightPermitBypass, r.Weight)
+		require.Equal(t, uint64(fenceWeightPermitBypass), r.Weight)
 		require.Len(t, r.Conditions, 1)
 		require.Equal(t, wf.FieldIPRemoteAddress, r.Conditions[0].Field)
 		require.Equal(t, want[r.Layer], r.Conditions[0].Value)
